Abstract engageUser() to TaskManager
